relayer: move AptosFunctionParams
